test: isolate allowance report helper coverage
